requirement.txt
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
ConfigArgParse==1.2.3
Flask==1.1.2
Flask-BasicAuth==0.2.0
gevent==20.6.2
geventhttpclient==1.4.4
greenlet==0.4.16
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
locust==1.2.3
MarkupSafe==1.1.1
msgpack==1.0.0
psutil==5.7.2
pycparser==2.20
pyzmq==19.0.2
requests==2.24.0
six==1.15.0
urllib3==1.25.10
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.1.0
